text-minimessage: Reduce the API surface of the parse context a little bit
